Fix compiler warnings about pointer-integer conversions
- Use the proper null constant (OBJ_NULL, ID_NULL, 0, ...) depending on
  the context
- When (mis)using a void pointer to hold an integer argument, always
  cast to the native integer type

+++ src/GameSrc/digifx.c
@@ -302,7 +302,7 @@ int play_digi_fx_master(int sfx_code, int num_loops, ObjID id, ushort x, ushort 
 #endif
     {
         if (id != OBJ_NULL)
-            s_dprm.data = (void *)id;
+            s_dprm.data = (void *)(unsigned int)id;
         else if (x != 0)
             s_dprm.data = (void *)(0x80000000 | (x << 16) | y);
         else
